Dockerize/requirements.txt
streamlit==1.3.1
matplotlib>=3.2.2
numpy==1.16.6
opencv-python>=4.1.2
tensorflow==1.14.0
PyYAML>=5.3.1
tensorflow-object-detection-api==0.1.1